clean up drawer and get it working properly, WIP
