+++ Commerce-Project/client/src/components/index.tsx
@@ -1,3 +1,2 @@
 export * from './NavBar';
 export * from './Theme';
-export * from './Drawer';
